DL-4387: Remove redundant routes
